updating prudhvi streams and creating StreamsEx5.java and StreamsEx6.java
