+++ .gitignore
@@ -1,4 +1,5 @@
 ### Hybris template
+out/
 /core-customize/README
 /Moon/README
 /Moon/.idea/
