+++ src/main/java/com/enh/streams/exercises/prudhvi/StreamsEx3.java
@@ -1,37 +1,34 @@
 package com.enh.streams.exercises.prudhvi;
 
 import com.enh.streams.StreamUtils;
-import com.enh.streams.data.Product;
 
-//        Reduction and Aggregation:
+import java.math.BigDecimal;
+import java.util.stream.Collectors;
+
 public class StreamsEx3 {
-    public static void main(String[] args) {
-//        a. Find the maximum value in a list of integers using streams.
-        System.out.println(
-                StreamUtils.getRandomIntegers().stream()
-                        .mapToInt(n -> n)
-                        .max()
-                        .orElse(0)
-        );
-        
-//        b. Calculate the product of all elements in a list of doubles.
-        System.out.println(
-                StreamUtils.getOrderedDoubles().stream()
-                        .reduce(1.0, (a, cv) -> a * cv)
-        );
-        
-//        c. Concatenate a list of strings using reduce operation.
-        System.out.println(
-                StreamUtils.getStrings().stream()
-                        .reduce("", (a, cs) -> a.concat(cs))
-        );
-        
-//        d. Count the number of occurrences of a specific word in a list of strings.
-        System.out.println(
-                StreamUtils.createProducts().stream()
-                        .map(Product::getDescription)
-                        .filter(d -> d.contains("galaxy"))
-                        .count()
-        );
-    }
+
+//    Reduction:
+public static void main(String[] args) {
+
+//    a. Given a list of integers, find and print the sum of all the elements.
+    System.out.println(
+            StreamUtils.getOrderedNumbers().stream()
+                    .mapToInt(i -> i).sum()
+    );
+
+//    b. Given a list of strings, concatenate them into a single string and print the result.
+    System.out.println(
+            StreamUtils.getStrings().stream()
+                    .reduce("",(p,c) -> p.concat(c))
+    );
+
+//    c. Given a list of objects, find the maximum or minimum value of a specific property.
+    System.out.println(
+            StreamUtils.createProducts().stream()
+                    .map(a -> a.getPrice())
+                    .min(BigDecimal::compareTo)
+                    .stream().collect(Collectors.toList())
+    );
+
+}
 }
