+++ src/main/java/com/enh/streams/exercises/prudhvi/StreamsEx6.java
@@ -1,34 +1,27 @@
 package com.enh.streams.exercises.prudhvi;
 
 import com.enh.streams.StreamUtils;
+import com.enh.streams.data.Employee;
 
-import java.math.BigDecimal;
 import java.util.stream.Collectors;
 
 public class StreamsEx6 {
-
-//    Reduction:
+//    Grouping and Partitioning:
 public static void main(String[] args) {
+//    a. Given a list of objects, group them by a specific property and print the groups.
+    System.out.println(
+            StreamUtils.createEmployees().stream()
+                    .collect(Collectors.groupingBy(Employee::getDepartments))
+    );
+/*    System.out.println("************");
+        StreamUtils.createEmployees().stream()
+            .collect(Collectors.groupingBy(Employee::getGender))
+            .forEach((d,e) -> e.forEach(System.out::println));*/
 
-//    a. Given a list of integers, find and print the sum of all the elements.
+//    b. Given a list of integers, partition them into two lists - even and odd.
     System.out.println(
             StreamUtils.getOrderedNumbers().stream()
-                    .mapToInt(i -> i).sum()
+                    .collect(Collectors.partitioningBy(i -> i % 2 == 0))
     );
-
-//    b. Given a list of strings, concatenate them into a single string and print the result.
-    System.out.println(
-            StreamUtils.getStrings().stream()
-                    .reduce("",(p,c) -> p.concat(c))
-    );
-
-//    c. Given a list of objects, find the maximum or minimum value of a specific property.
-    System.out.println(
-            StreamUtils.createProducts().stream()
-                    .map(a -> a.getPrice())
-                    .min(BigDecimal::compareTo)
-                    .stream().collect(Collectors.toList())
-    );
-
-}
+    }
 }
